Simplify what 'fell' exports
Fell was previously exporting every class that was used internally,
including what were essentially private utility classes, and was
exporting the log itself using a property name ('Log') that looked as
though it contained a class instead of an instance.

Instead, we now provide people with direct access to the log instance,
plus make the destination classes available so that they can configure
fell.