Allow code coverage to be generated locally
Although this isn't a CI coverage solution, and so doesn't have many of
the negatives of those solutions:

  1. Doesn't display coverage numbers against projects, encouraging
     people to write tests just to increase coverage.
  2. Doesn't show graphs of changing coverage over time, encouraging
     people to write tests just to increase coverage.
  3. Doesn't fail the build unless a minimum coverage target is met,
     encouraging people to write tests just to increase coverage.

It could still be misused by QAs that think it's a good idea to write
tests solely for the purpose of increasing coverage. Why this is a bad
idea is covered here:

  * http://www.dan-manges.com/blog/increasing-code-coverage-may-be-harmful

and even so, it doesn't even help developers to see whether the current
commits they are working on are properly covered, as envisaged by tools
like git-coverage:

  * http://stef.thewalter.net/git-coverage-useful-code-coverage.html

Maybe the answer is to have the command print a message like this once it's finished:

```sh
{red}With Great Power Comes Great Responsibility:{/red}

Adding tests solely for the purpose of increasing code coverage
considered harmful:

  * http://www.dan-manges.com/blog/increasing-code-coverage-may-be-harmful

Good tests verify the desired behavior of the program in the tersest,
most easy to comprehend manner. Your coverage report is available as
'coverage-report.html'.

{green}Please use this information in a responsible manner.{/green}
```